# openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /api/v1/menus:
    get:
      tags:
        - Menu CRUD
        - get
      summary: Get All Menus
      description: Get all menus
      operationId: get_all_menus_api_v1_menus_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/OutputMenuSchema'
                type: array
                title: Response Get All Menus Api V1 Menus Get
    post:
      tags:
        - Menu CRUD
        - post
      summary: Post Menu
      description: Create one menu
      operationId: post_menu_api_v1_menus_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuSchema'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputMenuSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}:
    get:
      tags:
        - Menu CRUD
        - get
      summary: Get Menu
      description: Get one menu by id
      operationId: get_menu_api_v1_menus__menu_id__get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputMenuSchema'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Menu CRUD
        - delete
      summary: Delete Menu
      description: Delete one menu
      operationId: delete_menu_api_v1_menus__menu_id__delete
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - Menu CRUD
        - patch
      summary: Update Menu
      description: Delete one menu
      operationId: update_menu_api_v1_menus__menu_id__patch
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputMenuSchema'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}/submenus:
    get:
      tags:
        - Submenu CRUD
        - get
      summary: Get All Submenus
      description: Get all submenus
      operationId: get_all_submenus_api_v1_menus__menu_id__submenus_get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutputSubmenuSchema'
                title: Response Get All Submenus Api V1 Menus  Menu Id  Submenus Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Submenu CRUD
        - post
      summary: Post Submenu
      description: Create one submenu
      operationId: post_submenu_api_v1_menus__menu_id__submenus_post
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmenuSchema'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputSubmenuSchema'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}/submenus/{submenu_id}:
    get:
      tags:
        - Submenu CRUD
        - get
      summary: Get Submenu
      description: Get one submenu by id
      operationId: get_submenu_api_v1_menus__menu_id__submenus__submenu_id__get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputSubmenuSchema'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Submenu CRUD
        - delete
      summary: Delete Submenu
      description: Delete one submenu
      operationId: delete_submenu_api_v1_menus__menu_id__submenus__submenu_id__delete
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - Submenu CRUD
        - patch
      summary: Update Submenu
      description: Delete one submenu
      operationId: update_submenu_api_v1_menus__menu_id__submenus__submenu_id__patch
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmenuSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputSubmenuSchema'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes:
    get:
      tags:
        - Dish CRUD
        - get
      summary: Get All Dishes
      description: Get all dishes
      operationId: get_all_dishes_api_v1_menus__menu_id__submenus__submenu_id__dishes_get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutputDishSchema'
                title: >-
                  Response Get All Dishes Api V1 Menus  Menu Id  Submenus
                  Submenu Id  Dishes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Dish CRUD
        - post
      summary: Post Dish
      description: Create one dish
      operationId: post_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes_post
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishSchema'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputDishSchema'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/menus/{menu_id}/submenus/{submenu_id}/dishes/{dish_id}:
    get:
      tags:
        - Dish CRUD
        - get
      summary: Get Dish
      description: Get one dish by id
      operationId: >-
        get_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__get
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
        - name: dish_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Dish Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputDishSchema'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Dish CRUD
        - delete
      summary: Delete Dish
      description: Delete one dish
      operationId: >-
        delete_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__delete
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
        - name: dish_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Dish Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - Dish CRUD
        - patch
      summary: Update Dish
      description: Delete one dish
      operationId: >-
        update_dish_api_v1_menus__menu_id__submenus__submenu_id__dishes__dish_id__patch
      parameters:
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Menu Id
        - name: submenu_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Submenu Id
        - name: dish_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Dish Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutputDishSchema'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionS'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    DishSchema:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        price:
          type: string
          title: Price
      type: object
      required:
        - title
        - description
        - price
      title: DishSchema
      description: Dish input schema
    ExceptionS:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
        - detail
      title: ExceptionS
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MenuSchema:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
      type: object
      required:
        - title
        - description
      title: MenuSchema
      description: Menu input schema
    OutputDishSchema:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        price:
          type: string
          title: Price
        id:
          type: string
          format: uuid
          title: Id
        submenu_id:
          type: string
          format: uuid
          title: Submenu Id
      type: object
      required:
        - title
        - description
        - price
        - id
        - submenu_id
      title: OutputDishSchema
      description: Dish output schema
    OutputMenuSchema:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        id:
          type: string
          format: uuid
          title: Id
        submenus_count:
          type: integer
          title: Submenus Count
        dishes_count:
          type: integer
          title: Dishes Count
      type: object
      required:
        - title
        - description
        - id
        - submenus_count
        - dishes_count
      title: OutputMenuSchema
      description: Menu output schema
    OutputSubmenuSchema:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        id:
          type: string
          format: uuid
          title: Id
        dishes_count:
          type: integer
          title: Dishes Count
        menu_id:
          type: string
          format: uuid
          title: Menu Id
      type: object
      required:
        - title
        - description
        - id
        - dishes_count
        - menu_id
      title: OutputSubmenuSchema
      description: Submenu output schema
    SubmenuSchema:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
      type: object
      required:
        - title
        - description
      title: SubmenuSchema
      description: Submenu input schema
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError